go can not recognize double correctly, so use int first

+++ include/goavx/cv/videoio.h
@@ -20,7 +20,7 @@ int _cv_videocapture_open_device(VideoCapturePtr cap, int device);
 void _cv_release_videocapture(VideoCapturePtr cap);
 int _cv_videocapture_open_file(VideoCapturePtr cap, const char* filename);
 bool _cv_videocapture_is_opened(VideoCapturePtr cap);
-double _cv_videocapture_get(VideoCapturePtr cap, int propId);
+int _cv_videocapture_get(VideoCapturePtr cap, int propId);
 
 #ifdef __cplusplus
 }
